backend/requirements.txt
# Flask and extensions
flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.0

# Database
pymongo==4.5.0

# Document processing
pypdf==3.15.1
PyPDF2==3.0.1
python-docx==0.8.11
pdfminer.six==20221105
python-magic==0.4.27
reportlab==4.1.0

# HTTP and API
requests==2.31.0
python-multipart==0.0.6
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# Utilities
python-dateutil==2.8.2
Werkzeug==2.3.7

# Knowledge Graph
networkx==3.1
numpy<2.0
scipy==1.11.4
matplotlib==3.7.2
